requirements.txt
-i https://pypi.org/simple
attrs==24.2.0; python_version >= '3.7'
aws-cdk-lib==2.162.1; python_version ~= '3.8'
aws-cdk.asset-awscli-v1==2.2.208; python_version ~= '3.8'
aws-cdk.asset-kubectl-v20==2.1.3; python_version ~= '3.8'
aws-cdk.asset-node-proxy-agent-v6==2.1.0; python_version ~= '3.8'
aws-cdk.cloud-assembly-schema==38.0.1; python_version ~= '3.8'
cattrs==24.1.2; python_version >= '3.8'
cdk-nag==2.29.12; python_version ~= '3.8'
constructs==10.4.2; python_version ~= '3.8'
importlib-resources==6.4.5; python_version >= '3.8'
jsii==1.104.0; python_version ~= '3.8'
publication==0.0.3
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typeguard==2.13.3; python_full_version >= '3.5.3'
typing-extensions==4.12.2; python_version >= '3.8'
